+++ stdnet/utils/jsontools.py
@@ -76,7 +76,10 @@ class JSONRPCEncoder(json.JSONEncoder):
 class jsonPickler(object):
     
     def dumps(self, obj, **kwargs):
-        return json.dumps(res, cls=JSONRPCEncoder, **kwargs)
+        return json.dumps(obj, cls=JSONDateDecimalEncoder, **kwargs)
     
     def loads(self,sobj):
         return json.loads(sobj)
+
+
+jsonpickler = jsonPickler()