+++ tests/profile/delete_t/tests.py
@@ -1,0 +1,11 @@
+from stdnet import test
+from stdnet.utils import populate, zip
+
+from profile.delete import tests as dtest
+
+
+class PipeDeleteTest(dtest.DeleteTest):                        
+    
+    def run(self):
+        self.model.objects.all().delete()
+